# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer DB API",
        "description": "Customer database API: authentication with access and refresh tokens, customer management, avatars, CSV import/export and background jobs.",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "v1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Create a staff account and start a session. The first administrator is created by the seeder.",
                "operationId": "2951b1c064a4f9057f7298ac6d8792a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Exchange email and password for an access token and an HttpOnly refresh cookie.",
                "operationId": "d16e6f2130a19a6802c1344ae5bcd3da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Signed in"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh session",
                "description": "Rotate the refresh cookie and issue a new access token. Reusing a rotated token revokes the session family.",
                "operationId": "920d0a9fa5c09f34b36431d472c48963",
                "responses": {
                    "200": {
                        "description": "Session refreshed"
                    },
                    "401": {
                        "description": "Missing, expired or reused refresh token"
                    }
                }
            }
        },
...
        "/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Current user",
                "description": "Current account profile.",
                "operationId": "ef26d2cc6c09ef674a2c765586a91bc9",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/auth/profile": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update profile",
                "description": "Update the display name of the current account.",
                "operationId": "aead5d28592d9ae3372f85226ac45447",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change password",
                "description": "Change the password and revoke every other active session.",
                "operationId": "eb29bfba9b5d3365299eaf0b504db90e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "new_password",
                                    "new_password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed"
                    },
                    "400": {
                        "description": "Validation failed"
                    },
                    "401": {
                        "description": "Missing or invalid token"
                    }
                },
                "security": [
                    {
                        "auth_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "auth_token": {
                "type": "apiKey",
                "description": "JWT issued by the application; send the raw token or 'Bearer <token>'.",
                "name": "Authorization",
                "in": "header"
            },
            "health_token": {
                "type": "apiKey",
                "description": "Token configured through HEALTHCHECK_TOKEN for detailed health diagnostics.",
                "name": "X-Health-Token",
                "in": "header"
            }
        }
    },
...
}